+++ experimental/src/test/kotlin/net/corda/contracts/universal/FXSwap.kt
@@ -1,6 +1,6 @@
 package net.corda.contracts.universal
 
-import net.corda.core.utilities.DUMMY_NOTARY
+import net.corda.testing.DUMMY_NOTARY
 import net.corda.testing.transaction
 import org.junit.Ignore
 import org.junit.Test