+++ node/src/test/kotlin/net/corda/node/services/persistence/DataVendingServiceTests.kt
@@ -12,7 +12,7 @@ import net.corda.core.flows.InitiatingFlow
 import net.corda.core.identity.Party
 import net.corda.core.node.services.unconsumedStates
 import net.corda.core.transactions.SignedTransaction
-import net.corda.core.utilities.DUMMY_NOTARY
+import net.corda.testing.DUMMY_NOTARY
 import net.corda.flows.BroadcastTransactionFlow.NotifyTxRequest
 import net.corda.node.services.NotifyTransactionHandler
 import net.corda.node.utilities.transaction