+++ test-utils/src/main/kotlin/net/corda/testing/driver/NetworkMapStartStrategy.kt
@@ -1,7 +1,7 @@
 package net.corda.testing.driver
 
 import com.google.common.net.HostAndPort
-import net.corda.core.utilities.DUMMY_MAP
+import net.corda.testing.DUMMY_MAP
 import org.bouncycastle.asn1.x500.X500Name
 
 sealed class NetworkMapStartStrategy {